add logic files for display, game logic and menu logic and remove then from main, update main
