+++ bin/game.rb
@@ -1,0 +1,6 @@
+require_relative 'menu.rb'
+
+module Game
+
+ 
+end
